Fix registry rebuild
Before this change, reloading the registry generated corrupted classes. This is because the original definition of class fragments to be assembled is modified during the build process. If this process was restarted, the definition would no longer be the original one, but the one containing the class hierarchy calculated during the previous build.
When building the registry, this problem is avoided by taking a copy of the original definition as the starting point for assembling the class.